[add] update/migration whit rollback

+++ database/migrations/2024_05_13_152209_create_trains_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->string('departure_time', 5);
             $table->string('Arrival time', 5);
             $table->smallInteger('train_code');
-            $table->smallInteger('number_of_carriages')->unsigned();
+            $table->tinyInteger('number_of_carriages')->unsigned();
             $table->boolean('in_time')->default(true);
             $table->boolean('deleted')->default(false);
             $table->timestamps();
